move to policy.go
