+++ resources/providers/awslib/iam/policy.go
@@ -19,14 +19,100 @@ package iam
 
 import (
 	"context"
+	"encoding/json"
+	"fmt"
 	iamsdk "github.com/aws/aws-sdk-go-v2/service/iam"
 	"github.com/aws/aws-sdk-go-v2/service/iam/types"
+	"github.com/elastic/cloudbeat/resources/fetching"
+	"github.com/elastic/cloudbeat/resources/providers/awslib"
+	"net/url"
 )
+
+func (p Provider) GetPolicies(ctx context.Context) ([]awslib.AwsResource, error) {
+	var policies []awslib.AwsResource
+
+	input := &iamsdk.ListPoliciesInput{OnlyAttached: true}
+	for {
+		listPoliciesOutput, err := p.client.ListPolicies(ctx, input)
+		if err != nil {
+			return nil, err
+		}
+
+		for _, policy := range listPoliciesOutput.Policies {
+			output, err := p.getPolicyVersion(ctx, policy)
+			if err != nil {
+				return nil, err
+			}
+
+			doc, err := decodePolicyDocument(output.PolicyVersion)
+			if err != nil {
+				return nil, err
+			}
+			policies = append(policies, Policy{
+				Policy:   policy,
+				Document: doc,
+			})
+		}
+
+		if !listPoliciesOutput.IsTruncated {
+			break
+		}
+		input.Marker = listPoliciesOutput.Marker
+	}
+
+	return policies, nil
+}
+
+func (p Provider) getPolicyVersion(ctx context.Context, policy types.Policy) (*iamsdk.GetPolicyVersionOutput, error) {
+	if policy.Arn == nil || policy.DefaultVersionId == nil {
+		return nil, fmt.Errorf("invalid policy: %v", policy)
+	}
+	return p.client.GetPolicyVersion(ctx, &iamsdk.GetPolicyVersionInput{PolicyArn: policy.Arn, VersionId: policy.DefaultVersionId})
+}
+
+func decodePolicyDocument(policyVersion *types.PolicyVersion) (map[string]interface{}, error) {
+	if policyVersion == nil || policyVersion.Document == nil {
+		return nil, fmt.Errorf("invalid policy version: %v", policyVersion)
+	}
+
+	// The policy document is URL-encoded, compliant with RFC 3986
+	docString, err := url.QueryUnescape(*policyVersion.Document)
+	if err != nil {
+		return nil, fmt.Errorf("failed to unescape policy document: %w", err)
+	}
+
+	var doc map[string]interface{}
+	err = json.Unmarshal([]byte(docString), &doc)
+	if err != nil {
+		return nil, fmt.Errorf("failed to unmarshal policy document: %w", err)
+	}
+
+	return doc, nil
+}
+
+func (p Policy) GetResourceArn() string {
+	return stringOrEmpty(p.Arn)
+}
+
+func (p Policy) GetResourceName() string {
+	return stringOrEmpty(p.PolicyName)
+}
+
+func (p Policy) GetResourceType() string {
+	return fetching.PolicyType
+}
+
+func stringOrEmpty(s *string) string {
+	if s == nil {
+		return ""
+	}
+	return *s
+}
 
 func (p Provider) listAttachedPolicies(ctx context.Context, identity *string) ([]types.AttachedPolicy, error) {
 	p.log.Debugf("listAttachedPolicies for user: %s", *identity)
 	input := &iamsdk.ListAttachedUserPoliciesInput{UserName: identity}
-	policies := []types.AttachedPolicy{}
+	var policies []types.AttachedPolicy
 	for {
 		output, err := p.client.ListAttachedUserPolicies(ctx, input)
 		if err != nil {
@@ -62,7 +148,7 @@ func (p Provider) listInlinePolicies(ctx context.Context, identity *string) ([]P
 		input.Marker = output.Marker
 	}
 
-	policies := []PolicyDocument{}
+	var policies []PolicyDocument
 	for i := range policyNames {
 		inlinePolicy, err := p.client.GetUserPolicy(ctx, &iamsdk.GetUserPolicyInput{
 			PolicyName: &policyNames[i],
